Change indent width to two spaces
The code looked to much shifted to the right in many places
(especially the specs), and the four-space indent resulted in
e.g. the require.js module indent to be overglorified. I have
the feeling that the code is more approachable this way and
one doesn't have to fight against too-long lines that much.