Introduce ModelBase, CollectionBase classes
These classes are now parent to all other types of models.
For now, they offer overrides of fetch(), save() and friends
which accept a UserCredentials object to make an "Authorization"
header of. This centralizes the several instances of this code
scattered around the codebase. Other functionality may move here
as well in the future.