Introduce User model, only show posting box when allowed to post
The User model is more or less a container for UserCredentials and
SubscribedChannels. Its main purpose is to make it easier to pass
these two models around without requiring both a "credentials" and
a "subscribedChannels" option to all interested views.

UserCredentials.verify() now moved to User.login() and tests the
credentials by directly fetching SubscribedChannels instead of
doing a wasted "GET /". register() should probably move to User
too in the future.