Make login work across tabs
The way this works is that the credentials are saved in localStorage
(in addition to sessionStorage) with a "loginCount", which acts like
a reference count: User#login() increases it by one, while the new
User#logout() does the opposite. If a tab is closed and the  count
drops to zero, the credentials are removed from localStorage.