Session handling fixes: - Not storing raw password anymore, storing basic auth header credentials instead; - Not using sessionStorage to handle temporary sessions, using session cookies instead; - Not using localStorage to handle permanent sessions, using expirable (7 days) cookies instead; - Tests fixes.
Now, when the user does not check 'store credentials', it's session will last until the browser is closed
(if is not configured to remember previous sessions). If the user selects 'store credentials', the session will
be active for 7 days after the last navigation action.

In the future, we are planning to add auth token based sessions, that means that we would have the option
to send a basic auth header or this specfic auth token when we are trying to validate an user action.